Improve Docker image ports and healthcheck
Expose DNS ports 5335/tcp and 5335/udp from the Docker image and add a HTTP healthcheck that polls the /health endpoint for a JSON status of 'ok'. Also introduce an FH_PRIORITY Makefile variable and propagate it as a com.foghorn.priority label on docker-run containers so multiple Foghorn instances can be ordered externally.